backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
eventlet==0.38.2
fastapi==0.115.6
fastapi-cli==0.0.7
Flask==3.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.0
fonttools==4.54.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.18.6
multidict==6.1.0
narwhals==1.13.3
numpy==2.1.3
oauthlib==3.2.2
openai==1.54.3
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
pyparsing==3.2.0
PyPDF2==3.0.1
pyphen==0.15.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.1
python-jose==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.12.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.21.0
rsa==4.9
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.40.0
tenacity==9.0.0
text-unidecode==1.3
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
# uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.0.3
weasyprint==63.0
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.17.1
zopfli==0.2.3.post1
pymongo==4.6.1
boto3==1.34.69
botocore==1.34.69
pytest==8.3.3
moto==4.2.14
mongomock==4.1.2
